2010-12-23  Lucas De Marchi  <lucas.demarchi@profusion.mobi>
        Reviewed by Kenneth Rohde Christiansen.

        [EFL] Clean warnings in Tiled Backing Store
        https://bugs.webkit.org/show_bug.cgi?id=51529

        Tiled backing store in EFL was giving the following warnings that are
        gone now:

        - Static function defined but not used;
        - Comparison between signed and unsigned integer expressions;

        * ewk/ewk_tiled_backing_store.c:
        (_ewk_tiled_backing_store_item_fill):
        (_ewk_tiled_backing_store_smart_calculate):
        * ewk/ewk_tiled_matrix.c:
        (_ewk_tile_matrix_slicer_setup):

Canonical link: https://commits.webkit.org/64893@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@74554 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebKit/efl/ewk/ewk_tiled_matrix.c
@@ -600,9 +600,9 @@ static Eina_Bool _ewk_tile_matrix_slicer_setup(Ewk_Tile_Matrix *tm, const Eina_R
         y = 0;
     }
 
-    if (y + h - 1 > rows * th)
+    if (y + h - 1 > (long)(rows * th))
         h = rows * th - y;
-    if (x + w - 1 > cols * tw)
+    if (x + w - 1 > (long)(cols * tw))
         w = cols * tw - x;
 
     return eina_tile_grid_slicer_setup(slicer, x, y, w, h, tw, th);
